Add support for lport parent_name and tag.
A logical port has an optional parent logical port.  This is for the
case of a container running inside of a VM.  The VM's port is set as a
parent of ports created for containers running in that VM.  Those
container ports also have a 'tag' that is used by OVN to differentiate
traffic to/from the VM itself and each container.

This patch adds the underlying support needed for setting these
fields.  The next patch will use it based on input provided in neutron
requests for creating or updating ports.

Change-Id: I9a9a9db340f68268a8f0b6fed4f1b9c6e3bbf246