Add port 'up' and 'down' notification
This adds the ability to watch for OVSDB events and take action
when they match specific conditions.

This patch handles the Logical_Port.up updates
 - At startup, it catches the initial 'create' events from
   ovsdb server and updates the port status if not in sync.
 - If the Logical_Port.up changes from False to True
   it sets the port status to 'ACTIVE' so that neutron server
   can send 'network-vif-plugged' event to nova.
 - If the Logical_Port.up changes from True to False
   it sets the port status to 'DOWN' so that neutron server
   can send 'network-vif-unplugged' event to nova.

To ensure that only one neutron-server handles the OVSDB events
it uses the ovsdb lock.

When the port is created, it sets the status to 'DOWN' by default.

Co-Authored-By: Numan Siddique <nusiddiq@redhat.com>

Change-Id: Ib7e1ebbbbd430a46aead2402e8b7543b5d73806e
Closes-bug: #1457571