Fix several ovsdb verify() issues.
When doing a read/modify/write operation, we must call verify() on a
row *before* modifying it.  This ensures that we don't complete any
writes based on dirty reads.

We had some cases where verify() was missing and others where verify()
was called *after* writing to a column, which turns the verify() call
into a no-op.

Change-Id: I616e72df7fbf17674f8739287126fb7f98b5347c
Closes-bug: #1544999
Signed-off-by: Russell Bryant <rbryant@redhat.com>