Reduce number of OVSDB row changes during ACL updates
This patch set reduces the number of ACL rows changed during the
ACL update algorithm. The patch compares the existing ACLs
associated with a port with the new set of ACLs. Only the difference
between the sets of ACLs are submitted to the OVS transaction list.

Improvement Seen:
-----------------

Consider a scenario where ports are attached to a network (provider
or private), and each port is assigned an unique IP Address i.e. a
common scenario, and is attached to the default security group.
Without the patch, a sample setup was seeing OVSDB transaction
timeouts when the 101th port was added to the network. This is
because the existing code deleted all (7) acls existing 100 ports and
re-added 6 acls back and modified 1 acl to include the 101th ports
ip address. This also resulted in (7+7)*100=1400 row changes in the
OVN_Northbound DB, and consequently 1400 notification of changes need
to be processed by processes monitoring changes in OVN_Northbound DB.

With the patch for the above case only (1 add +1 delete)*100 = 200
rows are modified and 200 row change notification sent to the
clients.

Result:
Existing Code: OVSDB transaction timeout seen in 101st port addition.
With the patch: OVSDB transaction timeout seen in the 270th port
addition - this timeout was not related to the 270*2=540 ACL
transactions, but seems to related to other OVSDB activity.
Regardless there is an improvement of 2.7X in terms of the number of
ports allowed to be provisioned in a provider or private network.

Change-Id: Ib968583685128ba44c91c30bcd522595645273da
Closes-bug: #1545189