Refactor ovsdb frontend code from neutron and networking-ovn
This patch reorganizes files for OVSDB operations from both neutron
and networking-ovn. As for Open vSwitch management, this includes
both a vsctl and a native IDL implementations; only the native
IDL implementation is instead available for OVN - an implementation
based on ovn-nbctl might be provided in the future, assuming there's
a real case for building it.

This patch also "imports" the unit tests for ovsdb_monitor from
networking-ovn. No other unit tests specific to ovsdb management was
found both in neutron and networking-ovn.

Files are reorganized as follows:

- oslo_ovsdb_frontend
    - api: API interface both for OVS and OVN
    - impl: Implementations: ovs_vsctl, ovs & ovn native
        - native: aux modules for the ovn native implementation

The code in this repo is structured, for the time being, as an
oslo library.